migrate from qemu to K210, to be continued
1. entry.S - add fix-mapping for low address after page table switch
2. init.c - temporarily comment cons_init()
	Origin cons_init() consists of a simple get_char() statement,
	which blocks the kernel when running on K210, so removing it
	is necessary. But it seems the K210 board cannot receive
	keyboard input now, which implies we have to do something
	initializing console rather than that on qemu put nothing
	in cons_init(). Now the shell doesn't work due to inability
	to receive keyboard input, but next commit will fix it.
3. memlayout.h - modify macros for k210 memory layout
        - KERNBASE 0xFFFFFFFFC0200000 => 0xFFFFFFFFC0020000
        - KMEMSIZE 0x7E00000 => 0x5E0000
        - PHYSICAL_MEMORY_END 0x88000000 => 0x80600000
        - KERNEL_BEGIN_PADDR 0x80200000 => 0x80020000
        - KERNEL_BEGIN_VADDR 0xFFFFFFFFC0200000 => 0xFFFFFFFFC0020000
4. pmm.c - adjust va_pa_offset & abandon static assert for KERNBASE alignment
5. proc.c - adjust type of kernel argc from int to long
	In the end of load_icode(), we prepare arguments for user main,
	which is achieved by an intermediate umain(). The arguments here,
	as convention, should be argc and argv. Argv
	consists of pointers of 8 bytes, while argc of type int is 4
	byte long, which means, stacktop won't align with 8 byte. After
	transferring control to userspace, ignorant of this misalignment,
	Risc-V R/W misalignment exception will be triggered. So during
	intermediate argument passing, we modify the type of umain argc
	from int to long and cast it back before jump into user's main,
	which keeps stack frame aligned to 8 byte without losing precision.
6. trap.c - cancel set for sstatus's SUM bit
        In v1.11 SUM permits S-mode to access userspace address, while
        in v1.9, the same bit is PUM, which on set will forbid such
        access. If wrongly set it in v1.11 spec for v1.9, later accesses
        to userspace adrress will trigger page fault
7. riscv.h - modify sstatus bit definition with respect to v1.9 rather than v1.11
        - change SUM to PUM
        - delete UXL
8. kernel.ld - start virtual address of kernel(BASE_ADDRESS) 0xFFFFFFFFC0200000 => 0xFFFFFFFFC0020000
9. user/libs/syscall.c - "lw" instrcutions before "ecall" are corrected to "ld"(consistent with uint64_t)

+++ lab8/user/libs/umain.c
@@ -20,7 +20,7 @@ initfd(int fd2, const char *path, uint32_t open_flags) {
 }
 
 void
-umain(int argc, char *argv[]) {
+umain(long argc, char *argv[]) {
     int fd;
     if ((fd = initfd(0, "stdin:", O_RDONLY)) < 0) {
         warn("open <stdin> failed: %e.\n", fd);
@@ -28,7 +28,7 @@ umain(int argc, char *argv[]) {
     if ((fd = initfd(1, "stdout:", O_WRONLY)) < 0) {
         warn("open <stdout> failed: %e.\n", fd);
     }
-    int ret = main(argc, argv);
+    int ret = main((int)argc, argv);
     exit(ret);
 }
